Refactored tests for readability. Also fixed delete method to take a long instead of int

+++ src/main/java/uk/co/davidatkins/simpleaccountservice/repo/AccountRepository.java
@@ -9,5 +9,5 @@ public interface AccountRepository {
 
     void addAccount(Account build);
 
-    void deleteAccount(int i);
+    void deleteAccount(long i);
 }
